# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/{userId}/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "hobbyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all transactions for a user",
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "summary": "Create a new transaction",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/users/{userId}/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "summary": "Get transaction by ID",
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "summary": "Update transaction by ID",
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Transaction deleted successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "summary": "Delete transaction by ID",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/transaction-categories": {
      "get": {
        "operationId": "TransactionCategoriesController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all transaction categories",
        "tags": [
          "Transaction Categories"
        ]
      },
      "post": {
        "operationId": "TransactionCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "summary": "Create transaction category",
        "tags": [
          "Transaction Categories"
        ]
      }
    },
    "/transaction-categories/{id}": {
      "get": {
        "operationId": "TransactionCategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "summary": "Get transaction category by ID",
        "tags": [
          "Transaction Categories"
        ]
      },
      "patch": {
        "operationId": "TransactionCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "summary": "Update transaction category",
        "tags": [
          "Transaction Categories"
        ]
      },
      "delete": {
        "operationId": "TransactionCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete transaction category",
        "tags": [
          "Transaction Categories"
        ]
      }
    },
    "/transaction-categories/{id}/tax-years/{taxYearId}": {
      "get": {
        "operationId": "TransactionCategoriesController_findMapping",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryMappingDto"
                }
              }
            }
          }
        },
        "summary": "Get category mapping for tax year",
        "tags": [
          "Transaction Categories"
        ]
      },
      "post": {
        "operationId": "TransactionCategoriesController_createMapping",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryMappingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryMappingDto"
                }
              }
            }
          }
        },
        "summary": "Create category mapping for tax year",
        "tags": [
          "Transaction Categories"
        ]
      },
      "patch": {
        "operationId": "TransactionCategoriesController_updateMapping",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryMappingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryMappingDto"
                }
              }
            }
          }
        },
        "summary": "Update category mapping for tax year",
        "tags": [
          "Transaction Categories"
        ]
      },
      "delete": {
        "operationId": "TransactionCategoriesController_removeMapping",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxYearId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete category mapping for tax year",
        "tags": [
          "Transaction Categories"
        ]
      }
    }
  },
  "info": {
    "title": "spndr-api",
    "description": "API for Spndr, a platform for managing and tracking financial flows of multiple income streams.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number"
              },
              "offset": {
                "type": "number"
              }
            },
            "required": [
              "total",
              "offset"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cln1234567890"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "amount": {
            "type": "string",
            "example": "199.99"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string",
            "example": "Website Development"
          },
          "reference": {
            "type": "string",
            "nullable": true,
            "example": "INV-2024-001"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "hobbyId": {
            "type": "string",
            "nullable": true,
            "example": "cln1234567890"
          },
          "categoryId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "taxYearId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "amount",
          "date",
          "description",
          "userId",
          "categoryId",
          "taxYearId",
          "createdAt",
          "updatedAt"
        ]
      },
      "i": {
        "type": "object",
        "properties": {}
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "amount": {
            "example": "199.99",
            "description": "Transaction amount",
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "date": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "description": {
            "type": "string",
            "example": "Website Development"
          },
          "reference": {
            "type": "string",
            "example": "INV-2024-001"
          },
          "notes": {
            "type": "string"
          },
          "hobbyId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "categoryId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "bankAccountId": {
            "type": "string",
            "example": "cln1234567890"
          }
        },
        "required": [
          "type",
          "amount",
          "date",
          "description",
          "categoryId"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "amount": {
            "example": "199.99",
            "description": "Transaction amount",
            "allOf": [
              {
                "$ref": "#/components/schemas/i"
              }
            ]
          },
          "date": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "description": {
            "type": "string",
            "example": "Website Development"
          },
          "reference": {
            "type": "string",
            "example": "INV-2024-001"
          },
          "notes": {
            "type": "string"
          },
          "hobbyId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "categoryId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "bankAccountId": {
            "type": "string",
            "example": "cln1234567890"
          }
        }
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cln1234567890"
          },
          "name": {
            "type": "string",
            "example": "Sales"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Sales/turnover"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sales"
          },
          "description": {
            "type": "string",
            "example": "Sales/turnover"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sales"
          },
          "description": {
            "type": "string",
            "example": "Sales/turnover"
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE"
            ]
          }
        }
      },
      "CategoryMappingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cln1234567890"
          },
          "categoryId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "taxYearId": {
            "type": "string",
            "example": "cln1234567890"
          },
          "sa103fBox": {
            "type": "string",
            "enum": [
              "BOX_15",
              "BOX_16",
              "BOX_17",
              "BOX_18",
              "BOX_19",
              "BOX_20",
              "BOX_21",
              "BOX_22",
              "BOX_23",
              "BOX_24",
              "BOX_25",
              "BOX_26",
              "BOX_27",
              "BOX_28",
              "BOX_29",
              "BOX_30",
              "BOX_31",
              "BOX_32",
              "BOX_33",
              "BOX_34",
              "BOX_35",
              "BOX_36",
              "BOX_37",
              "BOX_38",
              "BOX_39",
              "BOX_40",
              "BOX_41",
              "BOX_42",
              "BOX_43",
              "BOX_44",
              "BOX_45",
              "BOX_46"
            ]
          },
          "allowable": {
            "type": "boolean",
            "example": true
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Additional notes about tax treatment"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "categoryId",
          "taxYearId",
          "sa103fBox",
          "allowable",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCategoryMappingDto": {
        "type": "object",
        "properties": {
          "sa103fBox": {
            "type": "string",
            "enum": [
              "BOX_15",
              "BOX_16",
              "BOX_17",
              "BOX_18",
              "BOX_19",
              "BOX_20",
              "BOX_21",
              "BOX_22",
              "BOX_23",
              "BOX_24",
              "BOX_25",
              "BOX_26",
              "BOX_27",
              "BOX_28",
              "BOX_29",
              "BOX_30",
              "BOX_31",
              "BOX_32",
              "BOX_33",
              "BOX_34",
              "BOX_35",
              "BOX_36",
              "BOX_37",
              "BOX_38",
              "BOX_39",
              "BOX_40",
              "BOX_41",
              "BOX_42",
              "BOX_43",
              "BOX_44",
              "BOX_45",
              "BOX_46"
            ]
          },
          "allowable": {
            "type": "boolean",
            "example": true
          },
          "notes": {
            "type": "string",
            "example": "Additional notes about tax treatment"
          }
        },
        "required": [
          "sa103fBox",
          "allowable"
        ]
      },
      "UpdateCategoryMappingDto": {
        "type": "object",
        "properties": {
          "sa103fBox": {
            "type": "string",
            "enum": [
              "BOX_15",
              "BOX_16",
              "BOX_17",
              "BOX_18",
              "BOX_19",
              "BOX_20",
              "BOX_21",
              "BOX_22",
              "BOX_23",
              "BOX_24",
              "BOX_25",
              "BOX_26",
              "BOX_27",
              "BOX_28",
              "BOX_29",
              "BOX_30",
              "BOX_31",
              "BOX_32",
              "BOX_33",
              "BOX_34",
              "BOX_35",
              "BOX_36",
              "BOX_37",
              "BOX_38",
              "BOX_39",
              "BOX_40",
              "BOX_41",
              "BOX_42",
              "BOX_43",
              "BOX_44",
              "BOX_45",
              "BOX_46"
            ]
          },
          "allowable": {
            "type": "boolean",
            "example": true
          },
          "notes": {
            "type": "string",
            "example": "Additional notes about tax treatment"
          }
        }
      }
    }
  }
}